+++ YasEngineSoftware/Projectile.hpp
@@ -9,7 +9,7 @@ class Projectile : public YasPolygon
 		YasVector2D<float> velocity;
 		Projectile(float radius, float x, float y, YasVector2D<float> direction);
 		~Projectile();
-		void move(float deltaTime) override;
+		void move(double deltaTime) override;
 		void generate() override;
 		void generateRegularPolygonVertices(const YasVector2D<float>& position, float circumscribedCircleRadius, int numberOfVertices) override;
 		void regeneratePolygon() override;
